[CAN-58] feat: 노트 페이지 구분
