+++ src/components/note/noteDetail/SkeletonNoteModal.tsx
